+++ config/autoload/cache.global.php
@@ -6,7 +6,7 @@ use Shlinkio\Shlink\Core\Config\EnvVars;
 
 return (static function (): array {
     $redisServers = EnvVars::REDIS_SERVERS->loadFromEnv();
-    $redis = ['pub_sub_enabled' => $redisServers !== null && EnvVars::REDIS_PUB_SUB_ENABLED->loadFromEnv(false)];
+    $redis = ['pub_sub_enabled' => $redisServers !== null && EnvVars::REDIS_PUB_SUB_ENABLED->loadFromEnv()];
     $cacheRedisBlock = $redisServers === null ? [] : [
         'redis' => [
             'servers' => $redisServers,
@@ -16,7 +16,7 @@ return (static function (): array {
 
     return [
         'cache' => [
-            'namespace' => EnvVars::CACHE_NAMESPACE->loadFromEnv('Shlink'),
+            'namespace' => EnvVars::CACHE_NAMESPACE->loadFromEnv(),
             ...$cacheRedisBlock,
         ],
         'redis' => $redis,